新增API分组管理相关接口

+++ src/main/java/com/keqi/apihu/pj/domain/PjConstant.java
@@ -2,6 +2,9 @@ package com.keqi.apihu.pj.domain;
 
 public final class PjConstant {
 
-    public static final int rootParentId = 0;
+    /**
+     * API 分组顶级 parentId
+     */
+    public static final long rootParentId = 0;
 
 }
